requirements.txt
slither-analyzer==0.9.2
lxml==4.9.1
beautifulsoup4==4.11.1
requests==2.28.2
openai==0.27.4
pyyaml==6.0.1
solc-select==1.0.4
tiktoken==0.1.1